refactor: Update imports to use the new location of utility functions in the lib/utils directory
